move variant configuration reading in separate file
